requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
django==1.11.23
djangorestframework==3.9.4
pillow==3.4.2
pytz==2019.2              # via django
